change int to float

+++ 02-fahrenheit2celsius.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-	int fahrenheit, celsius;
-	int lower, upper, step;
+	float fahrenheit, celsius;
+	float lower, upper, step;
 
 	lower = 0;
 	upper = 300;
@@ -10,8 +10,8 @@ int main() {
 
 	fahrenheit = lower;
 	while (fahrenheit <= upper) {
-		celsius = 5 * (fahrenheit - 32) / 9;
-		printf("%3d %6d\n", fahrenheit, celsius);
+		celsius = 5.0 * (fahrenheit - 32.0) / 9.0;
+		printf("%3.0f %6.1f\n", fahrenheit, celsius);
 		fahrenheit += step;
 	}
 }
